fixed cmake static library

+++ src/client.cpp
@@ -2,9 +2,9 @@
 #include <iostream>
 #include <vector>
 
-#include <cheetah_interface.hpp>
-#include <hpmpc_interface.hpp>
-#include <networks/resnet50.hpp>
+#include <core/cheetah_interface.hpp>
+#include <core/hpmpc_interface.hpp>
+#include <core/networks/resnet50.hpp>
 
 #define PARTY 2
 
@@ -45,9 +45,10 @@ int main(int argc, char** argv) {
     {
         // num_triples = 30'000'000;
         num_triples = 1'168'448;
-        std::vector<uint64_t> a(num_triples, 1);
-        std::vector<uint64_t> b(num_triples, 1);
-        std::vector<uint64_t> c(num_triples, 1);
+        std::vector<uint32_t> a(num_triples, 1);
+        std::vector<uint32_t> b(num_triples, 1);
+        std::vector<uint32_t> c(num_triples, 1);
+
 
         Iface::generateArithTriplesCheetah(a.data(), b.data(), c.data(), 32, num_triples,
                                            std::string(addr), port, PARTY, threads);
